requirements.txt
langchain
langchain_community
streamlit
langchain_core
ollama
python-dotenv
langchain-ollama
